Conditional Sensor Input Update
Before this update, all sensors were turned on at all times. This was an issue because the car could no longer move if an object was detected.

With this update, the sensors are turned on based on the joystick input. For example, if the car is going forward, only the front sensors are turned on. This update allows the vehicle to move away from detected objects.